Converted all DTOs from using JsonProperty to SerializedName.
JsonProperty is used on getters/setters. I want my DTOs to be as clean as possible. This is why l use SerializedName tag which fits on top of the properties. JsonIgnore is put on getter. Expose is the GSON equivalent. I am using GSON to serialize JSONs so l need GSON tags.