Added public remote API controller.
ApiKey is loaded via VM options script. This is done to not have the apiKey exposed in the code. Sensitive data should be protected (like DB username and password).

This is the API l used in the start for my Android app. I will extract some data from it in JSON and load it in my DB in custom format that l choose.